+++ client/src/components/Chat.jsx
@@ -1,7 +1,37 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import ChatSelector from "./ChatSelector";
+import io from "socket.io-client";
+const socket = io("http://localhost:5000");
 
-const chat = () => {
+const Chat = () => {
+  const [message, setMessage] = useState("");
+  const [messageReceived, setMessageReceived] = useState("");
+  const [room, setRoom] = useState("");
+
+  const sendMessage = () => {
+    socket.emit("send_message", { message });
+  };
+
+  const joinRoom = () => {
+    if (room !== "") {
+      socket.emit("join_room", room);
+    }
+  };
+
+  //will always watch for recieve message event and alert the message
+  useEffect(() => {
+    socket.on(
+      "recieve_message",
+      (data) => {
+        setMessageReceived(data.message);
+      },
+      [socket]
+    );
+  });
+
+  //Need to implement room concept but use user icons to change room number.
+  //Need to implement the ability to generate a unique room number everytime a user is created.
+
   return (
     <section className="message-area">
       <div className="container">
@@ -72,6 +102,7 @@ const chat = () => {
                           >
                             {/* chat-list */}
                             <div className="chat-list">
+                              {/* stacked chats */}
                               <ChatSelector />
                             </div>
                             {/* chat-list */}
@@ -82,11 +113,9 @@ const chat = () => {
                             role="tabpanel"
                             aria-labelledby="Closed-tab"
                           >
-                            {/* chat-list */}
                             <div className="chat-list">
-                              <ChatSelector />
+                              Closed is now redundant (remove?)
                             </div>
-                            {/* chat-list */}
                           </div>
                         </div>
                       </div>
@@ -97,6 +126,7 @@ const chat = () => {
               </div>
               {/* chatlist */}
               {/* chatbox */}
+
               <div className="chatbox">
                 <div className="modal-dialog-scrollable">
                   <div className="modal-content">
@@ -119,11 +149,12 @@ const chat = () => {
                               />
                             </div>
                             <div className="flex-grow-1 ms-3">
-                              <h3>Person Name</h3>
+                              <h3>person name</h3>
                               <p>Item name</p>
                             </div>
                           </div>
                         </div>
+
                         <div className="col-4">
                           <ul className="moreoption">
                             <li className="navbar nav-item dropdown">
@@ -175,7 +206,7 @@ const chat = () => {
                             <p> Hey, Are you there? </p>
                             <span className="time">10:16 am</span>
                           </li>
-                          <li className="repaly">
+                          <li className="reply">
                             <p>yes!</p>
                             <span className="time">10:20 am</span>
                           </li>
@@ -187,7 +218,7 @@ const chat = () => {
                             <p> Hey, Are you there? </p>
                             <span className="time">10:32 am</span>
                           </li>
-                          <li className="repaly">
+                          <li className="reply">
                             <p>How are you?</p>
                             <span className="time">10:35 am</span>
                           </li>
@@ -196,13 +227,13 @@ const chat = () => {
                               <h6>Today</h6>
                             </div>
                           </li>
-                          <li className="repaly">
+                          <li className="reply">
                             <p> yes, tell me</p>
                             <span className="time">10:36 am</span>
                           </li>
-                          <li className="repaly">
-                            <p>yes... on it</p>
-                            <span className="time">junt now</span>
+                          <li className="reply">
+                            <p>{messageReceived}</p>
+                            <span className="time">Just Now</span>
                           </li>
                         </ul>
                       </div>
@@ -214,8 +245,11 @@ const chat = () => {
                           className="form-control"
                           aria-label="message…"
                           placeholder="Write message…"
+                          onChange={(event) => {
+                            setMessage(event.target.value);
+                          }}
                         />
-                        <button type="button">
+                        <button type="button" onClick={sendMessage}>
                           <i className="fa fa-paper-plane" aria-hidden="true" />{" "}
                           Send
                         </button>
@@ -255,4 +289,4 @@ const chat = () => {
   );
 };
 
-export default chat;
+export default Chat;